+++ CHANGELOG.md
@@ -602,3 +602,4 @@
 * Define `yio_yo` operator
 * Define `yolk` operator
 * Define `U_III_UU_I_II` data type
+* Define `W_II_II_I` data type